refactor: 더보기 설정 버튼 내부로직 사용자 훅으로 분리
